Changed the type of weight to `float` from `int`.

+++ src/load-balancer/src/Node.php
@@ -14,7 +14,7 @@ namespace Hyperf\LoadBalancer;
 class Node
 {
     /**
-     * @var int
+     * @var float
      */
     public $weight;
 
@@ -34,7 +34,7 @@ class Node
      */
     public $pathPrefix = '';
 
-    public function __construct(string $host, int $port, int $weight = 0, string $pathPrefix = '')
+    public function __construct(string $host, int $port, float $weight = 0, string $pathPrefix = '')
     {
         $this->host = $host;
         $this->port = $port;
